refactor: hydrate service interface
