+++ apps/renderer/src/services/entry.ts
@@ -1,14 +1,18 @@
 import { browserDB } from "~/database"
+import { appLog } from "~/lib/log"
 import type { EntryModel } from "~/models/types"
+import type { FlatEntryModel } from "~/store/entry"
+import { entryActions, useEntryStore } from "~/store/entry"
 
 import { BaseService } from "./base"
 import { CleanerService } from "./cleaner"
 import { EntryRelatedKey, EntryRelatedService } from "./entry-related"
+import type { Hydable } from "./interface"
 
 type EntryCollection = {
   createdAt: string
 }
-class EntryServiceStatic extends BaseService<EntryModel> {
+class EntryServiceStatic extends BaseService<EntryModel> implements Hydable {
   constructor() {
     super(browserDB.entries)
   }
@@ -77,6 +81,7 @@ class EntryServiceStatic extends BaseService<EntryModel> {
       this.table.bulkDelete(entryIds),
       EntryRelatedService.deleteItems(EntryRelatedKey.READ, entryIds),
       EntryRelatedService.deleteItems(EntryRelatedKey.COLLECTION, entryIds),
+      CleanerService.cleanRefById(entryIds),
     ])
   }
 
@@ -87,6 +92,48 @@ class EntryServiceStatic extends BaseService<EntryModel> {
       EntryRelatedService.deleteItems(EntryRelatedKey.READ, deleteEntryIds),
       EntryRelatedService.deleteItems(EntryRelatedKey.COLLECTION, deleteEntryIds),
     ])
+  }
+
+  async hydrate() {
+    const [entries, entryRelated, feedEntries, collections] = await Promise.all([
+      EntryService.findAll(),
+
+      EntryRelatedService.findAll(EntryRelatedKey.READ),
+      EntryRelatedService.findAll(EntryRelatedKey.FEED_ID),
+      EntryRelatedService.findAll(EntryRelatedKey.COLLECTION),
+    ])
+
+    const storeValue = [] as FlatEntryModel[]
+    const dirtyEntryIds = [] as string[]
+    for (const entry of entries) {
+      const entryRelatedFeedId = entry.feedId || feedEntries[entry.id]
+      if (!entryRelatedFeedId) {
+        appLog(`[Data hydrate warning]: Entry ${entry.id} has no related feed id`)
+        dirtyEntryIds.push(entry.id)
+        continue
+      }
+
+      storeValue.push({
+        entries: entry,
+        feedId: entryRelatedFeedId,
+        read: entryRelated[entry.id] || false,
+        collections: collections[entry.id] as {
+          createdAt: string
+        },
+      })
+    }
+    entryActions.hydrate(storeValue)
+    useEntryStore.setState({
+      starIds: new Set(Object.keys(collections)),
+    })
+
+    if (dirtyEntryIds.length > 0) {
+      // Remove entries that have no related feed id
+      EntryService.deleteEntries(dirtyEntryIds)
+      appLog(
+        `[Data hydrate warning]: Entry ${dirtyEntryIds.join(", ")} has no related feed id, cleanup..`,
+      )
+    }
   }
 }
