+++ apps/renderer/src/services/feed-unread.ts
@@ -1,7 +1,10 @@
 import { browserDB } from "~/database"
+import { feedUnreadActions } from "~/store/unread"
+
+import type { Hydable } from "./interface"
 
 const feedUnreadModel = browserDB.feedUnreads
-class ServiceStatic {
+class ServiceStatic implements Hydable {
   updateFeedUnread(list: [string, number][]) {
     return feedUnreadModel.bulkPut(list.map(([feedId, count]) => ({ id: feedId, count })))
   }
@@ -21,6 +24,12 @@ class ServiceStatic {
 
   async bulkDelete(ids: string[]) {
     return feedUnreadModel.bulkDelete(ids)
+  }
+
+  async hydrate() {
+    const unread = await FeedUnreadService.getAll()
+
+    return feedUnreadActions.hydrate(unread)
   }
 }
