+++ apps/renderer/src/services/interface.ts
@@ -1,0 +1,3 @@
+export interface Hydable {
+  hydrate: () => Promise<void>
+}
